--- validator.go
@@ -1,6 +1,0 @@
-package dotweb
-
-// Validator is the interface that wraps the Validate function.
-type Validator interface {
-	Validate(i interface{}) error
-}
